[JENKINS-23116] Split test classes that can work with @ClassRule and requires @Rule.
